% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded.R
\name{embedded_packages_are_great}
\alias{embedded_packages_are_great}
\title{embedded packages are great}
\usage{
embedded_packages_are_great()
}
\value{
-
}
\description{
-
}
\details{
-
}
\examples{
embedded_packages_are_great()
}
