% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_package}
\alias{lint_package}
\title{wrapper around lintr::lint_packages which throws error when code issues are detected}
\usage{
lint_package(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[lintr]{lint}}, e.g.
\code{cache} or \code{linters}.}
}
\value{
-
}
\description{
-
}
\details{
-
}
\seealso{
\code{\link[lintr]{lint_package}}
}
